Fix bug in index_to_pair (created in 727a1cb3fa756)

+++ g2o/core/base_constant_edge.h
@@ -51,7 +51,7 @@ namespace g2o {
 
     constexpr std::pair<int, int> index_to_pair(const int k)
     {
-      return {k - (ct_sqrt(2*(k+1))+.5) * (ct_sqrt(2*(k+1))+.5 - 1) / 2, ct_sqrt(2*(k+1))+.5};
+      return {k - static_cast<int>(ct_sqrt(2*(k+1))+.5) * (static_cast<int>(ct_sqrt(2*(k+1))+.5) - 1) / 2, ct_sqrt(2*(k+1))+.5};
     }
   }
